Bug 1849599 - Reorganize the Smoke tests suite
